+++ src/deployment-model/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { ComputeModelDescriptor } from './compute/compute-model-descriptor.js';
 import { EnvironmentStrategyDescriptor } from './environments/environment-strategy-descriptor.js';
 import { DeploymentOverviewDescriptor } from './overview/deployment-overview-descriptor.js';
 import { RuntimeTopologyDescriptor } from './runtime/runtime-topology-descriptor.js';
@@ -8,6 +9,7 @@ export function addDeploymentModel(services) {
     throw new Error('addDeploymentModel expects an instance of ServiceCollection.');
   }
 
+  services.registerSingleton('ComputeModelDescriptor', () => new ComputeModelDescriptor());
   services.registerSingleton('DeploymentOverviewDescriptor', () => new DeploymentOverviewDescriptor());
   services.registerSingleton('EnvironmentStrategyDescriptor', () => new EnvironmentStrategyDescriptor());
   services.registerSingleton('RuntimeTopologyDescriptor', () => new RuntimeTopologyDescriptor());